websocket/fastapi-websocket-tickets-auth/src/pyproject.toml
[tool.poetry]
name = "example-ws-service"
version = "0.1.0"
description = ""
authors = ["Ruslan Mamleev <mamleev.rus@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
redis-om = "^0.0.19"
websockets = "^10.1"
fastapi = "^0.73.0"
uvicorn = "^0.17.5"
orjson = "^3.6.7"
asyncio-redis = "^0.16.0"
backoff = "^1.11.1"
aiohttp = "^3.8.1"
loguru = "^0.6.0"
pytest = "^7.0.1"
requests = "^2.27.1"
pytest-xdist = "^2.5.0"
pytest-cov = "^3.0.0"

[tool.poetry.dev-dependencies]
black = "19.10b0"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.931"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
